refactor: split api handlers into packages
