+++ api/internal/models/models.go
@@ -1,4 +1,4 @@
-package main
+package models
 
 import (
 	"encoding/json"
